test wallet signing before setting preferences
