+++ .dev.vars
@@ -1,0 +1,1 @@
+KV_NAMESPACE_ID=7a6e9597aa6d4e708c3e100c63ab2881
